Created error/caching function for TSP trees. Should begin caching TSP files.

+++ modules/researchkit/timing/timing.cpp
@@ -75,11 +75,11 @@ bool Timer::tick() {
     return false;
 }
 
-size_t Timer::getTick() {
+unsigned long long Timer::getTick() {
     return _tick;
 }
 
-size_t Timer::getCycles() {
+unsigned long long Timer::getCycles() {
     return _cycles;
 }
